fix: update translation handling and improve message formatting across commands

+++ events/on_guild_remove.py
@@ -1,6 +1,7 @@
 from utils import remove_guild_from_db
 from discord.ext import commands
 import shutil
+import os
 
 class OnGuildRemove(commands.Cog):
     def __init__(self, client):
@@ -8,7 +9,7 @@ class OnGuildRemove(commands.Cog):
 
     @commands.Cog.listener()
     async def on_guild_remove(self, guild):
-        shutil.rmtree(f'server_configs/{guild.id}')
+        if os.path.exists(f'server_configs/{guild.id}'): shutil.rmtree(f'server_configs/{guild.id}')
         await remove_guild_from_db(guild_id=guild.id)
 
 async def setup(client):
